requirements.txt
pydantic>=2.6.1
httpx>=0.28.1
google-generativeai>=0.3.0
Pillow>=10.0.0
reportlab>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
